requirements.txt
Flask
pandas
requests
gunicorn
google-cloud-storage
google-cloud-bigquery
google-auth
google-auth-oauthlib
google-auth-httplib2
oauth2client
beautifulsoup4
xlrd
cloudscraper
pdfkit
google-api-python-client
gspread
Authlib
Flask-Login
Flask-Session
google-api-core
xds-protos
gunicorn